fix: 초기 커리큘럼 캔버스 관련 코드 삭제
